02_canSniffer_GUI/requirements.txt
PySerial
PyQt5
PyQt5-tools
qtmodern